+++ src/Test/java/seedu/sophia/SophiaTest.java
@@ -1,0 +1,18 @@
+package seedu.sophia;  //same package as the class being tested
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class SophiaTest {
+    @Test
+    public void TestStorage(){
+        //Sophia sophia
+        assertEquals(2, 2);
+    }
+
+    @Test
+    public void anotherDummyTest(){
+        assertEquals(4, 4);
+    }
+}
