--- src/main/java/Storage.java
@@ -1,58 +1,0 @@
-import java.io.*;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
-
-public class Storage {
-    private final String file_path;
-
-    public Storage(String file_path) {
-        this.file_path = file_path;
-    }
-
-    public void save(TaskList tasklist) throws IOException {
-        File file = new File(file_path);
-        File parentDirectory = file.getParentFile();
-        if (!parentDirectory.exists()) {
-            parentDirectory.mkdir();
-        }
-        if (!file.exists()) {
-            file.createNewFile();
-        }
-        FileWriter fw = new FileWriter(file);
-        BufferedWriter bw = new BufferedWriter(fw);
-        tasklist.write(bw);
-        bw.close();
-    }
-
-    public List<Task> load() throws FileNotFoundException {
-        List<Task> tasks = new ArrayList<>();
-        File file = new File(file_path);
-        try {
-            BufferedReader reader = new BufferedReader(new FileReader(file));
-            reader.lines().forEach(line -> {
-                String[] t = line.split("\\|");
-                for(int i = 0; i < t.length; i++) { t[i] = t[i].trim(); }
-                //System.out.println(Arrays.toString(t));
-                if(Objects.equals(t[0], "T")) {
-                    Task temp = new TodoTask(t[2]);
-                    temp.setDone(t[1].equals("1"));
-                    tasks.add(temp);
-                } else if(Objects.equals(t[0], "D")) {
-                    Task temp = new DeadlineTask(t[2],t[3]);
-                    temp.setDone(t[1].equals("1"));
-                    tasks.add(temp);
-                } else if(Objects.equals(t[0], "E")){
-                    Task temp = new EventTask(t[2],t[3],t[4]);
-                    temp.setDone(t[1].equals("1"));
-                    tasks.add(temp);
-                }
-            });
-            //System.out.println(tasks.toString());
-        } catch( FileNotFoundException e) {
-            throw e;
-        }
-        return tasks;
-    }
-}
