+++ src/main/java/sophia/EventTask.java
@@ -1,3 +1,5 @@
+package sophia;
+
 import java.io.BufferedWriter;
 import java.io.IOException;
