+++ src/main/java/sophia/SophiaException.java
@@ -1,3 +1,5 @@
+package sophia;
+
 public class SophiaException extends Exception {
     public SophiaException(String message) {
         super(message);
